requirements.txt
Flask==2.3.3
numpy==1.25.2
opencv-python==4.8.0.76
Pillow==10.0.1
